Refactor the documentation structure #55
The documentation structure was inconsistent and unstructured. This commit
structures the documentation based on components and unities of meanings. All
units should be independent from each-other and should not include the words
`rucio` and `documentation` in the title, since this is ambiguous.

The following structure will be enforced:

- Getting started:
  All information about the `Rucio` project and concepts.

- Client:
  The documentation related to the rucio `clients`. This includes the
  installation and usage.

- Server:
  All information to set up and maintain a rucio server.

- Developer:
  All information for developers. This includes setting up the demo environment
  and contributing.

- ...

This commit also deleats the `Try Rucio!` page, since it does not contain any
usefull information.