Add a link checker to the CI/CD
The internet is highly volatile. Links can get outdated, thus the user of the
documentation would not find the correct site.

This commit adds a cron job to the CI/CD, which runs a
`markdown-link-checker`. This `markdown-link-checker` traverses all links in the
markdown files and calls them. If the return value is not okay, the link is
marked as not reachable and the job fails. This is a good indicator, that a link
is not present anymore.

It also fixed all broken external links.