Change the automatic file generation
The file generation for the auto generated ones had some problems. It:

- Was difficult to follow
- Had a lot of edge cases
- Created files with other owners in the `docs` folder
- Not centralized in this repository

This commit refactors the automatic file generation. In particular, it runs the
generation in a pre-build docker container and moves the generated files to an
quarantined directory. The file ownership gets changed there and all the
neccesary files get copied to their respective location.