Enforce the Python Client Api Classes to end with Client #55
There are files with multiple classes in the client module. We are
taking the first class as the name of the clients endpoint. This was
proven to be not suitable, since the client class is not always the
first one.

This commit only selects the first class, which ends in "client". This
should be sufficient enough.