Automatically generate the release notes #119
The release notes are duplicated in the GitHub release notes. We can
easily fetch them and convert them to markdown.

This is not expencive, since the GitHub Api call is cheap and the
release notes are not big. The script finished in under 5 seconds, which
makes it suitable to run on every build.