Make the rest_api_doc_spec.yaml file available as asset
The `rest_api_doc_spec.yaml` file is auto-generated from the rucio
project and is used to create the Rest Api Docs. Making the spec file
available allows users to access and reference it directly, instead of
creating it on their own. This allows them to e.g. use different
front-ends or analyze the file.