Update Rucio Image Tag (Hardcoded)
Change the downloaded rucio image from `fedora32-python3.8` to
`alma9-python3.9` because the former was outdated and caused errors in
the docs build process.

Note that this hardcodes the image tag and the tag needs to be changed
each time rucio is updated to reflect changes made there.

A "better" solution would be to query the "latest" tag, but this does
not exist on the github container repository (GHCR), to the best of my
knowledge.