Document environmental variables affecting the client #171
This commit adds a new page to the ‘User’ section. It should be expanded
in the future (e.g. the different mechanisms of configuration and how to
craft a minimal configuration file.

The environmental variables related to BitTorrent and Traces are merely
mentioned, without further description.

The ‘SITE’ environmental variable used by ‘lib/rucio/common/pcache.py’
is not mentioned because it is likely an error.